b2b-integrated/sprintplanner-python-fastmcp-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.3
python-multipart>=0.0.6
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.23
alembic>=1.12.1
fastmcp>=0.3.0
